requirements.txt
Django==5.0
Pillow==10.1.0
python-decouple==3.8
django-js-asset==2.2.0
django-mptt==0.15.0
django-allauth~=0.60.0